ui redesign

+++ termalpinterd/services/startUpService.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
 using termalpinterd.Interfaces;
 
 namespace termalpinterd.services
@@ -11,24 +12,28 @@ namespace termalpinterd.services
     public class StartUpService:IStartUpService
     {
         private const string AppName = "TermalPrinterApp";
-        public  void SetStartup(bool enable)
+        public void SetStartup(bool enable)
         {
-            string appPath = Application.ExecutablePath; // Ruta del ejecutable
+            string appPath = Path.GetFullPath(Application.ExecutablePath); // Obtener ruta absoluta
 
             using (RegistryKey registryKey = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Run", true))
             {
+                if (registryKey == null)
+                {
+                    MessageBox.Show("Error: No se pudo abrir la clave del Registro.");
+                    return;
+                }
+
+            
                 if (enable)
                 {
-                    registryKey.SetValue(AppName, $"\"{appPath}\"");
-                    
+                    registryKey.SetValue(AppName, $"\"{appPath}\""); // Se agregan comillas dobles
                 }
                 else
                 {
                     registryKey.DeleteValue(AppName, false);
                 }
             }
-
-            
         }
         // Método para comprobar si la app está en el inicio de Windows
         public bool IsStartupEnabled()
